+++ classes/BrConsole.php
@@ -30,6 +30,18 @@ class BrConsole extends BrSingleton {
 
   }
 
+  public function blue($message) {
+
+    return chr(27) . '[34m' . $message . chr(27) . '[0m';
+
+  }
+
+  public function purple($message) {
+
+    return chr(27) . '[35m' . $message . chr(27) . '[0m';
+
+  }
+
   public function __call($name, $arguments) {
 
     if ($arguments) {